+++ api/db/affiliations.js
@@ -67,6 +67,16 @@ const getPopulatedAffiliationsByStateId = ({
   return getAffiliationsByStateId_({ stateId, status, isAdmin });
 };
 
+const getAffiliationsByUserId = (userId, {db = knex } = {}) => {
+  return db('auth_affiliations')
+    .select(selectedColumns)
+    .where('auth_affiliations.user_id', userId )
+    .leftJoin('auth_roles', 'auth_affiliations.role_id', 'auth_roles.id')
+    .leftJoin('okta_users', 'auth_affiliations.user_id', 'okta_users.user_id')
+
+
+};
+
 const getAffiliationById = ({ stateId, affiliationId, db = knex }) => {
   return db('auth_affiliations')
     .select(selectedColumns)
@@ -157,5 +167,6 @@ module.exports = {
   getAllAffiliations,
   reduceAffiliations,
   getAllPopulatedAffiliations,
+  getAffiliationsByUserId,
   selectedColumns
 };