+++ api/routes/affiliations/get.js
@@ -3,8 +3,9 @@ const {
   getPopulatedAffiliationsByStateId: _getPopulatedAffiliationsByStateId,
   getPopulatedAffiliationById: _getPopulatedAffiliationsById,
   getAllPopulatedAffiliations: _getAllPopulatedAffiliations,
+  getAffiliationsByUserId: _getAffiliationsByUserId,
 } = require('../../db');
-const { can, validForState } = require('../../middleware');
+const { loggedIn, can, validForState } = require('../../middleware');
 
 module.exports = (
   app,
@@ -12,6 +13,7 @@ module.exports = (
     getPopulatedAffiliationsByStateId = _getPopulatedAffiliationsByStateId,
     getPopulatedAffiliationById = _getPopulatedAffiliationsById,
     getAllPopulatedAffiliations = _getAllPopulatedAffiliations,
+    getAffiliationsByUserId = _getAffiliationsByUserId,
   } = {}
 ) => {
   app.get(
@@ -79,5 +81,22 @@ module.exports = (
       }
     }
   );
+
+  app.get(
+    '/affiliations/me',
+    loggedIn,
+    async (request, response, next) => {
+      logger.info({
+        id: request.id,
+        message: `handling GET /me endpoint}`
+      });
+      try {
+        const resp = await getAffiliationsByUserId(request.user.id)
+        return response.send(resp)
+      }
+      catch(e){
+        return next(e)
+      }
+    })
 
 };